fix: avoid relying on case-sensitive filesystems
