--- src/Colors.rs
@@ -1,148 +1,0 @@
-#![allow(non_upper_case_globals)]
-pub const Antiquewhite: [u8; 3] = [250, 235, 215];
-pub const Aliceblue: [u8; 3] = [240, 248, 255];
-pub const Aqua: [u8; 3] = [0, 255, 255];
-pub const Aquamarine: [u8; 3] = [127, 255, 212];
-pub const Azure: [u8; 3] = [240, 255, 255];
-pub const Beige: [u8; 3] = [245, 245, 220];
-pub const Bisque: [u8; 3] = [255, 228, 196];
-pub const Black: [u8; 3] = [0, 0, 0];
-pub const Blanchedalmond: [u8; 3] = [255, 235, 205];
-pub const Blue: [u8; 3] = [0, 0, 255];
-pub const Blueviolet: [u8; 3] = [138, 43, 226];
-pub const Brown: [u8; 3] = [165, 42, 42];
-pub const Burlywood: [u8; 3] = [222, 184, 135];
-pub const Cadetblue: [u8; 3] = [95, 158, 160];
-pub const Chartreuse: [u8; 3] = [127, 255, 0];
-pub const Chocolate: [u8; 3] = [210, 105, 30];
-pub const Coral: [u8; 3] = [255, 127, 80];
-pub const Cornflowerblue: [u8; 3] = [100, 149, 237];
-pub const Cornsilk: [u8; 3] = [255, 248, 220];
-pub const Crimson: [u8; 3] = [220, 20, 60];
-pub const Cyan: [u8; 3] = [0, 255, 255];
-pub const Darkblue: [u8; 3] = [0, 0, 139];
-pub const Darkcyan: [u8; 3] = [0, 139, 139];
-pub const Darkgoldenrod: [u8; 3] = [184, 134, 11];
-pub const Darkgray: [u8; 3] = [169, 169, 169];
-pub const Darkgreen: [u8; 3] = [0, 100, 0];
-pub const Darkgrey: [u8; 3] = [169, 169, 169];
-pub const Darkkhaki: [u8; 3] = [189, 183, 107];
-pub const Darkmagenta: [u8; 3] = [139, 0, 139];
-pub const Darkolivegreen: [u8; 3] = [85, 107, 47];
-pub const Darkorange: [u8; 3] = [255, 140, 0];
-pub const Darkorchid: [u8; 3] = [153, 50, 204];
-pub const Darkred: [u8; 3] = [139, 0, 0];
-pub const Darksalmon: [u8; 3] = [233, 150, 122];
-pub const Darkseagreen: [u8; 3] = [143, 188, 143];
-pub const Darkslateblue: [u8; 3] = [72, 61, 139];
-pub const Darkslategray: [u8; 3] = [47, 79, 79];
-pub const Darkslategrey: [u8; 3] = [47, 79, 79];
-pub const Darkturquoise: [u8; 3] = [0, 206, 209];
-pub const Darkviolet: [u8; 3] = [148, 0, 211];
-pub const Deeppink: [u8; 3] = [255, 20, 147];
-pub const Deepskyblue: [u8; 3] = [0, 191, 255];
-pub const Dimgray: [u8; 3] = [105, 105, 105];
-pub const Dimgrey: [u8; 3] = [105, 105, 105];
-pub const Dodgerblue: [u8; 3] = [30, 144, 255];
-pub const Firebrick: [u8; 3] = [178, 34, 34];
-pub const Floralwhite: [u8; 3] = [255, 250, 240];
-pub const Forestgreen: [u8; 3] = [34, 139, 34];
-pub const Fuchsia: [u8; 3] = [255, 0, 255];
-pub const Gainsboro: [u8; 3] = [220, 220, 220];
-pub const Ghostwhite: [u8; 3] = [248, 248, 255];
-pub const Gold: [u8; 3] = [255, 215, 0];
-pub const Goldenrod: [u8; 3] = [218, 165, 32];
-pub const Gray: [u8; 3] = [128, 128, 128];
-pub const Green: [u8; 3] = [0, 128, 0];
-pub const Greenyellow: [u8; 3] = [173, 255, 47];
-pub const Grey: [u8; 3] = [128, 128, 128];
-pub const Honeydew: [u8; 3] = [240, 255, 240];
-pub const Hotpink: [u8; 3] = [255, 105, 180];
-pub const Indianred: [u8; 3] = [205, 92, 92];
-pub const Indigo: [u8; 3] = [75, 0, 130];
-pub const Ivory: [u8; 3] = [255, 255, 240];
-pub const Khaki: [u8; 3] = [240, 230, 140];
-pub const Lavender: [u8; 3] = [230, 230, 250];
-pub const Lavenderblush: [u8; 3] = [255, 240, 245];
-pub const Lawngreen: [u8; 3] = [124, 252, 0];
-pub const Lemonchiffon: [u8; 3] = [255, 250, 205];
-pub const Lightblue: [u8; 3] = [173, 216, 230];
-pub const Lightcoral: [u8; 3] = [240, 128, 128];
-pub const Lightcyan: [u8; 3] = [224, 255, 255];
-pub const Lightgoldenrodyellow: [u8; 3] = [250, 250, 210];
-pub const Lightgray: [u8; 3] = [211, 211, 211];
-pub const Lightgreen: [u8; 3] = [144, 238, 144];
-pub const Lightgrey: [u8; 3] = [211, 211, 211];
-pub const Lightpink: [u8; 3] = [255, 182, 193];
-pub const Lightsalmon: [u8; 3] = [255, 160, 122];
-pub const Lightseagreen: [u8; 3] = [32, 178, 170];
-pub const Lightskyblue: [u8; 3] = [135, 206, 250];
-pub const Lightslategray: [u8; 3] = [119, 136, 153];
-pub const Lightslategrey: [u8; 3] = [119, 136, 153];
-pub const Lightsteelblue: [u8; 3] = [176, 196, 222];
-pub const Lightyellow: [u8; 3] = [255, 255, 224];
-pub const Lime: [u8; 3] = [0, 255, 0];
-pub const Limegreen: [u8; 3] = [50, 205, 50];
-pub const Linen: [u8; 3] = [250, 240, 230];
-pub const Magenta: [u8; 3] = [255, 0, 255];
-pub const Maroon: [u8; 3] = [128, 0, 0];
-pub const Mediumaquamarine: [u8; 3] = [102, 205, 170];
-pub const Mediumblue: [u8; 3] = [0, 0, 205];
-pub const Mediumorchid: [u8; 3] = [186, 85, 211];
-pub const Mediumpurple: [u8; 3] = [147, 112, 219];
-pub const Mediumseagreen: [u8; 3] = [60, 179, 113];
-pub const Mediumslateblue: [u8; 3] = [123, 104, 238];
-pub const Mediumspringgreen: [u8; 3] = [0, 250, 154];
-pub const Mediumturquoise: [u8; 3] = [72, 209, 204];
-pub const Mediumvioletred: [u8; 3] = [199, 21, 133];
-pub const Midnightblue: [u8; 3] = [25, 25, 112];
-pub const Mintcream: [u8; 3] = [245, 255, 250];
-pub const Mistyrose: [u8; 3] = [255, 228, 225];
-pub const Moccasin: [u8; 3] = [255, 228, 181];
-pub const Navajowhite: [u8; 3] = [255, 222, 173];
-pub const Navy: [u8; 3] = [0, 0, 128];
-pub const Oldlace: [u8; 3] = [253, 245, 230];
-pub const Olive: [u8; 3] = [128, 128, 0];
-pub const Olivedrab: [u8; 3] = [107, 142, 35];
-pub const Orange: [u8; 3] = [255, 165, 0];
-pub const Orangered: [u8; 3] = [255, 69, 0];
-pub const Orchid: [u8; 3] = [218, 112, 214];
-pub const Palegoldenrod: [u8; 3] = [238, 232, 170];
-pub const Palegreen: [u8; 3] = [152, 251, 152];
-pub const Paleturquoise: [u8; 3] = [175, 238, 238];
-pub const Palevioletred: [u8; 3] = [219, 112, 147];
-pub const Papayawhip: [u8; 3] = [255, 239, 213];
-pub const Peachpuff: [u8; 3] = [255, 218, 185];
-pub const Peru: [u8; 3] = [205, 133, 63];
-pub const Pink: [u8; 3] = [255, 192, 203];
-pub const Plum: [u8; 3] = [221, 160, 221];
-pub const Powderblue: [u8; 3] = [176, 224, 230];
-pub const Purple: [u8; 3] = [128, 0, 128];
-pub const Rebeccapurple: [u8; 3] = [102, 51, 153];
-pub const Red: [u8; 3] = [255, 0, 0];
-pub const Rosybrown: [u8; 3] = [188, 143, 143];
-pub const Royalblue: [u8; 3] = [65, 105, 225];
-pub const Saddlebrown: [u8; 3] = [139, 69, 19];
-pub const Salmon: [u8; 3] = [250, 128, 114];
-pub const Sandybrown: [u8; 3] = [244, 164, 96];
-pub const Seagreen: [u8; 3] = [46, 139, 87];
-pub const Seashell: [u8; 3] = [255, 245, 238];
-pub const Sienna: [u8; 3] = [160, 82, 45];
-pub const Silver: [u8; 3] = [192, 192, 192];
-pub const Skyblue: [u8; 3] = [135, 206, 235];
-pub const Slateblue: [u8; 3] = [106, 90, 205];
-pub const Slategray: [u8; 3] = [112, 128, 144];
-pub const Slategrey: [u8; 3] = [112, 128, 144];
-pub const Snow: [u8; 3] = [255, 250, 250];
-pub const Springgreen: [u8; 3] = [0, 255, 127];
-pub const Steelblue: [u8; 3] = [70, 130, 180];
-pub const Tan: [u8; 3] = [210, 180, 140];
-pub const Teal: [u8; 3] = [0, 128, 128];
-pub const Thistle: [u8; 3] = [216, 191, 216];
-pub const Tomato: [u8; 3] = [255, 99, 71];
-pub const Turquoise: [u8; 3] = [64, 224, 208];
-pub const Violet: [u8; 3] = [238, 130, 238];
-pub const Wheat: [u8; 3] = [245, 222, 179];
-pub const White: [u8; 3] = [255, 255, 255];
-pub const Whitesmoke: [u8; 3] = [245, 245, 245];
-pub const Yellow: [u8; 3] = [255, 255, 0];
